Adding initial SLUG controller implementation
This change adds the initial implementation of the SLUG service
controller. It parses REST URLs for user and group information
and then routes the request to the appropriate object controller.
This code is untested and does not yet work, it is entirely a
work-in-progress. It does not load data from an external source,
it uses hardcoded test data as a shim for now.