Updating SLUG controllers to working status
This change updates the SLUG controller classes and overall
CherryPy configuration, fixing a few obvious and subtle bugs,
resulting in a basic working version of the SLUG service.

To run, simply:

$ python slug.py

Use a normal web browser, or Python requests, to interact with
the REST API and see the current, hard-coded test data.

Access and error logs are also now stored in /var/log/slug.