Add unit tests covering the main application script
This change adds a unit test suite that tests and exercises the
different components provided in the main application script,
including argument utilities and the main application routine.
Extensive CherryPy mocking is involved. The main invocation used
to launch the application at the end of the script is trivial
and ignored.