SLUGS - Release 1.0.0
This release contains the initial release of the SLUGS library.
SLUGS provides a basic, web-based authentication system that can
be used to verify user existence and user/group information. It
is currently intended to be used in a local environment and should
not be used in place of a robust, secure authentication system in
production.